ENH: allow for different fitting methods
Alternative methods can be set either on the command line (positional
argument "-f") or in the method file (using the variable "fitmethod").

The default behavior remains unchanged, if nothing is specified the
"leastsq" method will be used. Many of the fit statistics and estimates
for uncertainties in parameters are only performed for this method.
Therefore, the  user will be warned if the last fitting method in the
procedure is different from "leastsq".