ENH: Make all parameters and parameter constraints explicit
All parameters are now declared in the `params` list and their associated constraints (if any) managed by the `lmfit` module. For example, all kinetics rate constants are now explicits. This means that their calculated values are now available in the `parameters.fit` file and their related constraints shown in the `constraints.fit` file. See the discussion in #35.