Add the commands "pick_cest" and "plot_param"
"pick_cest" and "plot_param" existed as separate tools. This commit
makes them full "chemex" commands with their own parser. In addition,
their UI changed: "pick_cest" reads a standard CEST "experiment.cfg"
file with the option "-e" and a path where writing the output files
with the option "-o"; "plot_param" uses the option "-p" to read the
"parameters.fit" file and the option "-n" to get the name of the
parameter to display as a funtion of residue number.