Fix "selection" bug
The selection based on the command line option "-r" was done before the
creation of the parameters. This was at the origin of a bug when the
selection in the method file was including more profiles than the ones
that were first selected. Changing the order of the parameter creations
and profile selection solved the issue. Fix issue #6.