Fix the duplicate profile name bug
When two profiles had the same name in the list of profiles in the
experiment.toml file, one of the profile was omitted. This commit fix
this issue. Profiles are not overwritten anymore as long as they come
from two different files – as it should.

In addition, I updated the CEST_13C_LABEL_CN to make apparent the name
of the proton associated to the carbon. This is not per se necessary,
but it helps identifying the profile when plotted.

Fix issue #5.