Put constrained parameters in their own file
Split the writing of the non free parameters into "constrained.toml" and
"fixed.toml" files containing the constrained and fixed parameters of
the fit, respectively. Constrained are now explicitly written in the
constrained.toml file next to the value of each parameter. The
constraints.toml file is therefore unnecessary and has been removed.
Meanwhile the free.toml file was renamed fitted.toml to be more
consistent with the other file names.