Split "-r"/"selection" to "include" and "exclude"
This commit rename the "-r" command line option (to select profiles to
include in the fit) to "--include". It introduces a new command line
option called "--exclude" to exclude profiles from the fit.

In addition, the "SELECTION =" option in the method.toml files is renamed
"INCLUDE =". The "EXCLUDE =" option is also introduced to remove
profiles from the fit.

Re issue #22.