Add R1(Nz) and R1(2HzNz) experiments
Regular relaxation experiment type is introduced. R1(Nz) and R1(2HzNz)
are implemented, so that decays can be directly used in ChemEx instead
of just setting (and fixing) the R1 values. This was usually done for
estimating 1HN spin flip rates or correcting for the time spend along z
in CPMG experiments.

Two examples are provided in CPMG15N_IP and
CPMG_15N_TR.