Add model free parameter fit
This commit adds the possibility to use the model free parameters in the
fit to set the different relaxation rates. Right now it uses one
correlation time (tauc), one order parameter (s2) to account for
"infinitely" fast local motion. In addition, it includes the solvent
exchange rate of amide proton (khh) per residue. For now, it should work
well with NH groups.

!!! Note that "tauc" values are given in nanoseconds !!!

To take into account for other protons in the molecule, the new
condition "deuterated" is introduced. This should be set in the
[conditions] section of experiment.toml files. In the future, this will
probably converted into a "label" option that could also contain
information about other labels like "13C" and "15N".

To avoid redundancy the [MODEL_FREE] section in parameter.toml files has
removed. Instead the user can set the parameters "TAUC_A", "S2_A" and
"KHH_A" for state A (same for the other states if needed) exactly like
one would to set any type of parameters. This works even though the
model free is not used as the fitting model. In that case the model free
parameters are used to set up the starting values.