Add the "label" option in the "conditions" section
The "deuterated" option has been replaced by the "label" option. This
new entry is a list that describe the labeling scheme used for the
sample. In particular, the "2H" label replaces the "deuterated" option
and the "13C" option replaces the "cn_label" option used in some
experiments.

Example for a perdeuterated uniformly 13C-labeled sample:
:
[conditions]
label = ["2H", "13C"]