Merged in example/2st_binding (pull request #19)
Add example for "2st_binding" kinetic model

An example illustrating the use of the "2st_binding" kinetic model is
added. It is based on 15N CPMG and CEST data recorded on sample of 15N
C-ter Artemis and various concentration of unlabeled DBD Ligase IV.

The model was slightly tweaked, so that global fitting parameters are
the off-rate, "KOFF", and the dissociation constant, "KD".