Make "d2o" a fitting parameter in hd exch models
A new parameter "D2O" is introduced, so that the amount of D2O in the
sample can be fitted. It depends on the condition "d2o" set in the
experiment.toml file, which means that experiments recorded with
different amount of D2O can be analysed conjointly. The starting value
of the "D2O" parameters is equal to their associated "d2o" condition.

Re issue #27.