Merged in feature/add_coscest_1hn_ip_ap (pull request #20)
Add the "coscest_1hn_ip_ap" experiment

This is the cos-CEST version of the "cest_1hn_ip_ap" experiment. The
cosine modulated pulse is defined by the number of irradiation
frequency, "cos_n", and the number of points to simulate the shape,
"cos_res".

An example based on the data from the JMR paper (Yuwen, Bouvignies and
Kay, 2018) can be found in "examples/Experiments/COSCEST_1HN_IP_AP".