Add DCEST experiments to 'pick_cest'
This commit add the possibility to load multiple CEST and D-CEST
experiments with the 'pick_cest' module. When D-CEST experiments are
loaded a parameter called 'Number of SWs', corresponding to the number
of D-CEST spectral width that should be plotted. The default value is
3.0. This value can be any positive float number.

An example use of the 'pick_cest' module with D-CEST experiments is
given in the 'DCEST_15N' example folder.

Fix issue #38