Fix a bug with the '--mc' option
The MC methods of the different experiments add a problem of mismatched
number of points between the simulated and experimental data sets. This
commit fixes this issue.

In addition, this commit also extends the fix made on the '--bs' option
for the CEST experiments in the previous commit to the other experiment
types.

Fixes issue #39.
Re issue #41.