Add a nucleus-specific bootstrap analysis option
A new option "--bsn N" is introduced to run a nucleus-specific bootstrap
analysis. In this analysis, nuclei are randomly selected in the
bootstrap manner. That is, if we have a dataset that depends on the
nuclei {G2N, H8N, R9N, R9H}, potential new datasets could include the
profiles of the following sets of nuclei {H8N, H8N, R9N, R9H} or {G2N,
G2N, G2N, R9H}. When a nucleus appears multiple times, the associated
profiles are also included multiple times in the bootstrap sample.

Note that this way of running the bootstrap analysis can produce
datasets with different number of data points in them. For examples, if
profiles depending on H8N appears in multiple experiments and the ones
depending on G2N in only one, then the two datasets mentioned above
would have different number of data points. This goes against the main
principle underlying bootstrap analysis that normally requires that all
the newly sampled datasets are of the same size.

In addition, this commit introduces a modification of the protocol
for all these simulations (MC, bootstrap,...) in that now the same steps
as in the main calculations are run. Thus, each simulation reporduces
all the steps of the main calculation and the results are similarly
written on the disk.