Change how parameter statuses are set
This commit modifies how parameter statuses (fixed, fitted and
constrained) are set in the method.toml-like files.

Parameters can be set to be fitted (fixed) using the key "FIT" ("FIX")
and added to the associated list.

For example:
FIT = ["PB", "KEX_AB", "DW_AB, NUC->45N, B0->800.0MHz"]

Parameters can be constrained using mathematical expression via the key
"CONSTRAINTS". Constraints are given in the form of a list of
constraints. Parameters within the expression should be put into
brackets.

For example:
CONSTRAINTS = [
    "[R2_B] = [R2_A]",
    "[R1_B, B0->800.0MHz] = [R1_A, B0->800.0MHz]",
]

Keys are read in that order: "CONSTRAINTS" -> "FIX" -> "FIT"

In addition, this commit merges all the parameter.toml-like files in the
examples into a single parameters.toml file.