Add grid search and chi2 plotting options
This commit add the possibility to run an nD grid search and plot the
resulting chi2 values as 1D and 2D plots. Grid search can be run using
the new option "GRID" in the method.toml file. The grid is defined on a
parameters basis. The parameters defining the nD grid are fixed to the
value of the grid that is evaluated, while the other parameters are set
as defined by the "FIX", "FIT" and "CONSTRAINTS" options. Points can be
defined using a linear scale, a log scale or point by point:
 - Linear scale: "[PB] = lin(min, max, nb of points)"
 - Log scale: "[PB] = log(min, max, nb of points)"
 - Point by point: "[PB] = (0.01, 0.02, 0.04, 0.06)"

Example:
GRID = [
    "[PB] = log(0.03, 0.1, 10)",
    "[KEX_AB] = log(200.0, 1000.0, 10)",
    "[DW_AB, NUC->G43N] = lin(0.0, 10.0, 10)",
]

Parameters are selected as usual. For example, "[R2_A]" would select
all the de R2_A. "[R2_A, NUC->G43N]" would select all the R2_A values
of the nucleus G43N (multiple values if there are several temperatures,
B0 field).

At the end of the grid search the best point is selected and the
corresponding parameters are then used in the next step of the fitting
procedure. The chi2 values are reported in the grid.toml file as well as
in the form of 1D and 2D plots. For an nD grid with n > 1, multiple 1D
grids are plotted, one per parameter, corresponding to the parameter
values on the x axis and the minimum chi2 value along the other
dimension. Similarly, if n > 2, then a series of 2D chi2 surface
corresponding to each pair of parameters is plotted. 2D surface
correspond to 2D projection in which the minimum value in the other
dimension is taken.