Enhance grid search to deal with local parameters
This is a significant change on how grid search is run when parameters
that are independent from each others are involved in the grid. In this
case there is no need to calculate the grid points between these
parameters. The new algorithm starts by calculating the minimal
individual grids separately and then combines them together to recover
the global minimum and to plot 1D chi2 curve for each parameter and 2D
plots when possible.

The algorithm is much faster in the sense that a search involving 2 global
parameters and 10 independent residue specific parameters does not
generate a 12-dimensional grid, but 12 3-dimensional grids, which of
course is much faster and importantly retains all the information.
All the resulting grids are written in separate files. Plots are as
before for global parameters, meaning that for all the parameters that
are not plotted the values that give the minimum chi2 for each point of
the grid are selected. For local parameters, the best global parameter
values are used for the whole grid.