Fix a bug with 'grouping' algorithm
The algorithm for making groups of data, each group depending on
disjoint sets of parameters, was flawed. It depended on the order
profiles were presented to the function. The algorithm has been fully
rewritten making immune to the order of the profiles. In addition, the
use of set() which does not preserved order has limited to avoid any
undesired shuffling during the grid search.