Add the correction of the profile names
Each experiment depends on specific spin systems. ChemEx expected the
name of the profiles to be fully compatible with the spin system expected
by the experiment. For example, the antiphase 1H CPMG experiment
expected a name like G3H-N to run correctly because the experiment
depends on the IS spin system where I=1H and S=15N. This commit adds a
correction to the name to change it so that it complies to the
requirement.

For example, if G3N is the name of the AP 1H profile, it will be
corrected to G3H-N.

In addition, it is now possible to gives names using the 3-letter
convention for the amino acid type, like GLY3H-N.

Also, in order to be fully consistent between different experiments, the
nucleus name 'HN' is changed to just 'H'.