Update the Monte Carlo-like simulations
Monte Carlo, bootstrap and nucleus-specific bootstrap analysis are now
run when the option "STATISTICS" is set in the method file. This means
that these simulations can be run or not in each step of the method
file.

The syntax is the following:

[STEP1]
STATISTICS = {"MC"= 100}

where MC is the type of simulation and 100 is the number of simulations.

Types can be:
  - "MC" for Monte Carlo
  - "BS" for bootstrap
  - "BSN" for nucleus-specific bootstrap

To run two or more types of simulation just add additional pairs of
value:

[STEP1]
STATISTICS = {"MC"= 100, "BS"= 100}

The output for each kind of simulation is stored in a single file stored
in the directory corresponding to the step of the method.

Parameter values are stored in different columns. When no values are
available, which can be the case for nucleus-specific bootstrap
analysis, the characters "--" are used to fill the space.

Note that the command line options "--mc", "--bs" and "--bsn" have been
removed.