Plot all the possible ncycs for CPMG experiments
All the possible ncycs between 1 and the largest from the profile are
plotted. When odd ncyc are not advised to be recorded only the odd ncyc
from the profiles are plotted (and all the even ncyc...).

As of today, only one experiment is not designed to record odd ncyc:
cpmg_ch3_13c_h2c.