Refactor the code related to the kinetic models
A new plugin architecture has been implemented to facilitate the
addition of new kinetic models. The code for calculating some of the
constrained rates and populations have been changed to use some
optimization calculation instead of an analytical formula (when
possible and when the analytical formula is more complicated to read).

In addition, 'new' kinetic models have been added:
- "3st_linear": A <-> B <-> C
- "3st_fork": C <-> A <-> B
- "3st_triangle": identical to "3st"