Add models using temperature coefficients
Chemical shifts are now temperature-dependant. In addition, in
temperature dependant models, such as `2st_eyring`, it is now possible
to use temperature coefficient by adding the extension `.tc` to the
model name, eg `2st_eyring.tc`.

When temperature coefficients are used, the following relationship
holds:

`DW_AB = DWM_AB * T + DWP_AB`

where `DWM_AB` is the temperature coefficient in ppm/ºC, `DWP_AB` is the
chemical shift difference between A and B at 0ºC in ppm, and `T` is the
temperature in Celsius. `DWM_AB` and `DWP_AB` are fitted and
temperature-independant.