Update binding kinetic models with Kd,app
Binding kinetic models with multiple bound states have now an "apparent Kd" that corresponds to the Kd where all the bound states are counted as only one state. It is called "kd_app". The Kd that corresponds to the binding between the free and the "first" bound state is still called "kd_ab" if "a" is free and "b" is the "first" bound state.