Add the `cs_evolution_prior` flag to CEST and CPMG experiments (#281)
The `cs_evolution_prior` flag allows for simulations where the CEST or CPMG block occurs after chemical shift evolution. When this flag is set to true in the `[experiment]` section of the `experiment.toml` file, the initial magnetization is populated only for the observed state (`observed_state`), while all other terms are set to zero. This modification ensures accurate boundary conditions by reflecting the magnetization distribution prior to the exchange block.